+++ quickshell/settings/categories/ThemePane.qml
@@ -1,5 +1,6 @@
 import QtQuick
 import QtQuick.Layouts
+import QtQuick.Controls
 import "../../ui" as Ui
 import "../.."
 
@@ -21,6 +22,7 @@ Item {
         contentWidth: width
         contentHeight: col.implicitHeight
         clip: true
+        ScrollBar.vertical: Ui.ScrollBar {}
         boundsBehavior: Flickable.StopAtBounds
 
         ColumnLayout {